requirements.txt
faiss-cpu
sentence-transformers
pandas
fastapi
uvicorn
streamlit
requests
